rename mapinsets to insets
