+++ microapps/MapviewInsetsApp/app/proguard-rules.pro
