+++ microapps/MapviewInsetsApp/app/src/main/res/xml/data_extraction_rules.xml
